+++ crates/rpc-types-mev/src/eth_calls.rs
@@ -1,4 +1,5 @@
-use crate::u256_numeric_string;
+use crate::{u256_numeric_string, Privacy, Validity};
+
 use alloy_eips::BlockNumberOrTag;
 use alloy_primitives::{Address, Bytes, B256, U256};
 use serde::{Deserialize, Serialize};
@@ -82,6 +83,98 @@ pub struct EthCallBundleTransactionResult {
     /// Contains the return data if the transaction reverted
     #[serde(skip_serializing_if = "Option::is_none")]
     pub revert: Option<Bytes>,
+}
+
+/// Request for `eth_cancelBundle`
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct CancelBundleRequest {
+    /// Bundle hash of the bundle to be canceled
+    pub bundle_hash: String,
+}
+
+/// Request for `eth_cancelPrivateTransaction`
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct CancelPrivateTransactionRequest {
+    /// Transaction hash of the transaction to be canceled
+    pub tx_hash: B256,
+}
+
+/// Bundle of transactions for `eth_sendBundle`
+///
+/// Note: this is for `eth_sendBundle` and not `mev_sendBundle`
+///
+/// <https://docs.flashbots.net/flashbots-auction/searchers/advanced/rpc-endpoint#eth_sendbundle>
+#[derive(Default, Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub(crate) struct EthSendBundle {
+    /// A list of hex-encoded signed transactions
+    pub(crate) txs: Vec<Bytes>,
+    /// hex-encoded block number for which this bundle is valid
+    #[serde(with = "alloy_serde::quantity")]
+    pub(crate) block_number: u64,
+    /// unix timestamp when this bundle becomes active
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
+    pub(crate) min_timestamp: Option<u64>,
+    /// unix timestamp how long this bundle stays valid
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
+    pub(crate) max_timestamp: Option<u64>,
+    /// list of hashes of possibly reverting txs
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub(crate) reverting_tx_hashes: Vec<B256>,
+    /// UUID that can be used to cancel/replace this bundle
+    #[serde(default, rename = "replacementUuid", skip_serializing_if = "Option::is_none")]
+    pub(crate) replacement_uuid: Option<String>,
+}
+
+/// Response from the matchmaker after sending a bundle.
+#[derive(Deserialize, Debug, Serialize, Clone, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub(crate) struct EthBundleHash {
+    /// Hash of the bundle bodies.
+    pub(crate) bundle_hash: B256,
+}
+
+/// Response from the matchmaker after sending a bundle.
+#[derive(Deserialize, Debug, Serialize, Clone, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct SendBundleResponse {
+    /// Hash of the bundle bodies.
+    pub bundle_hash: B256,
+}
+
+/// Request for `eth_sendPrivateTransaction`
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct PrivateTransactionRequest {
+    /// raw signed transaction
+    pub tx: Bytes,
+    /// Hex-encoded number string, optional. Highest block number in which the transaction should
+    /// be included.
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
+    pub max_block_number: Option<u64>,
+    /// Preferences for private transaction.
+    #[serde(default, skip_serializing_if = "PrivateTransactionPreferences::is_empty")]
+    pub preferences: PrivateTransactionPreferences,
+}
+
+/// Additional preferences for `eth_sendPrivateTransaction`
+#[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq, Eq)]
+pub struct PrivateTransactionPreferences {
+    /// Requirements for the bundle to be included in the block.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub validity: Option<Validity>,
+    /// Preferences on what data should be shared about the bundle and its transactions
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub privacy: Option<Privacy>,
+}
+
+impl PrivateTransactionPreferences {
+    /// Returns true if the preferences are empty.
+    pub const fn is_empty(&self) -> bool {
+        self.validity.is_none() && self.privacy.is_none()
+    }
 }
 
 #[cfg(test)]
