+++ crates/rpc-types-mev/src/lib.rs
@@ -6,35 +6,18 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-// serde helper to serialize/deserialize u256 as numeric string
-mod eth_call_bundle;
-pub use eth_call_bundle::*;
+mod eth_calls;
+pub use eth_calls::*;
 
-mod eth_cancel_bundle;
-pub use eth_cancel_bundle::*;
+mod mev_calls;
+pub use mev_calls::*;
 
-mod eth_cancel_private_transaction;
-pub use eth_cancel_private_transaction::*;
-
-mod eth_send_bundle;
-pub use eth_send_bundle::*;
-
-mod eth_send_private_transaction;
-pub use eth_send_private_transaction::*;
-
-mod mev_send_bundle;
-pub use mev_send_bundle::*;
-
-mod mev_sim_bundle;
-pub use mev_sim_bundle::*;
-
-mod user_stats;
-pub use user_stats::*;
-
-mod bundle_stats;
-pub use bundle_stats::*;
+// types for stats endpoint like flashbots_getUserStats and flashbots_getBundleStats
+mod stats;
+pub use stats::*;
 
 mod common;
 pub use common::*;
 
+// serde helper to serialize/deserialize u256 as numeric string
 mod u256_numeric_string;
